updating the phenom device driver and renaming the device
