+++ alab_control/auto_sem_eds/__init__.py
@@ -1,1 +1,1 @@
-from .auto_sem_eds import SEMDevice, SEMState
+from .auto_sem_eds import SEMDevice, SEMError, InstrumentMode, OperationalMode, ImagingDevice
